refactor : landing 컴포넌트 언어 변경 JSX -> TSX
