+++ src/components/Landing/LandingSectionFour.tsx
@@ -1,19 +1,14 @@
-import React from 'react';
-import CalenderIcon from '@/assets/LandingPage-calender.webp';
+import Gallery from '@/assets/landing/landing-gallery.webp';
 
-function LandingPagePartTwo() {
+function LandingPagePartFour() {
   return (
     <div>
       <div className="flex h-[5.625rem] justify-around bg-gray-100/30 lg:w-[38.75rem] 2xl:w-[46.125rem]">
-        <img
-          src={CalenderIcon}
-          alt="plane"
-          className="h-[5.625rem] w-[5.625rem]"
-        />
+        <img src={Gallery} alt="plane" className="h-[5.625rem] w-[5.625rem]" />
 
         <div className="relative flex flex-col justify-center py-[1.6875rem]  pl-[1.723125rem] font-semibold text-contentsPrimary">
-          <p>리뷰를 확인하고</p>
-          <p>일정을 조율해 보세요!</p>
+          <p>다녀온 여행을</p>
+          <p>추억으로 남겨보세요!</p>
           <div
             className="absolute
 bottom-6  h-2 w-[8.775rem] bg-custom-color/40"
@@ -24,4 +19,4 @@ bottom-6  h-2 w-[8.775rem] bg-custom-color/40"
   );
 }
 
-export default LandingPagePartTwo;
+export default LandingPagePartFour;
